requirements.txt
selenium>=4.0.0
pyyaml>=5.4
webdriver-manager>=3.8.0
flask>=2.3.0
